Update readme.md
I added customer information to the trainingSessions table so you can see how to handle multiple training sessions for a single request.  (Many of the columns can be deleted after the various values are determined.  Things like date, time, and duration are only there to make it easier to generate scheduling emails.)